requirements.txt
requests==2.32.5
python-dotenv==1.2.1
Pillow==12.1.1
sv_ttk==2.6.1

